feat: add nav drop down to ui toolkit
