+++ libs/ui-toolkit/src/components/index.ts
@@ -19,6 +19,7 @@ export * from './link';
 export * from './loader';
 export * from './lozenge';
 export * from './nav';
+export * from './nav-dropdown';
 export * from './popover';
 export * from './price-change';
 export * from './progress-bar';
